backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
bcrypt==4.2.0
pyjwt==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.12
